0.5: RHEL support and stuff
- New boxes are now being used which both support
`config.vm.synced_folder`, which means that once you download a Monitor
file once, it will now stick around in /vagrant. This improves the speed
for re-deployment of an identical version. Remove the file if you don't
want to re-use it.

- RHEL 7 is now supported and requires you to fill out `secret.sh`
correctly.

- The `-m` flag in `Vagrantfile` for choosing a version seems to work
well and should now probably be used for most cases.

- Missing `-p` flag should now correctly not install PHP debugging
tools.

- README has been updated, log an issue if anything is unclear.